requirements.txt
git+https://github.com/marin-m/vmlinux-to-elf
ripgrepy
packaging
reindent
gitpython
rich